small checks and corrections to the logging system
I am doing that just to make sure the code follows the schematics of the project.
The most intrusive thing that I am planning to do is to change the email with username because that is the thing that we promised in the tasks and we have to deliver what we promised.