Please ignore the previous commit
My last commit that "fixing the login" was accidental. i was working on another project and I accidentally did I commited the logs of the intellj with the comment that I wanted to leave to the other poject.

I also fixed bit the lines in this commit but it is not enough and I will need to see how to improve that.