using the consumption rate
no the person, will consume each item according to the consumption rate of its product.
The catch here: if the consumption rate is lower that 1 then it uses propability instead of adding. That is more natural, and also it works better because we don't have to deliver the items fractioned.